+++ src/main/java/io/xlorey/FluxLoader/client/core/Core.java
@@ -1,5 +1,7 @@
 package io.xlorey.FluxLoader.client.core;
 
+import io.xlorey.FluxLoader.client.events.ClientEventsHandler;
+import io.xlorey.FluxLoader.shared.EventManager;
 import io.xlorey.FluxLoader.shared.PluginManager;
 import io.xlorey.FluxLoader.utils.Logger;
 
@@ -15,6 +17,8 @@ public class Core {
         Logger.printCredits();
 
         Logger.print("FluxLoader Core initialization for the client....");
+
+        EventManager.subscribe(new ClientEventsHandler());
 
         PluginManager.loadPlugins(true);
     }