+++ src/main/java/io/xlorey/FluxLoader/client/events/ClientEventsHandler.java
@@ -1,0 +1,7 @@
+package io.xlorey.FluxLoader.client.events;
+
+/**
+ * Client event handler
+ */
+public class ClientEventsHandler {
+}